Changed the old not semantic name to labelled flag
