+++ src/app/DTT/flag/index.js
@@ -1,12 +1,10 @@
-import { Box, HStack, Switch, Text } from '@chakra-ui/react'
-import { useEffect, useState } from 'react'
+import { Switch } from '@chakra-ui/react'
 
-import { apiConfig } from 'config/get-api-config'
-import { equals } from 'ramda'
-import { isNullOrUndefinedOrEmpty } from 'utils/helpers/is-null-or-undefined'
 import { onSendMessage } from 'vertx'
 import { selectCode } from 'redux/db/selectors'
 import { useSelector } from 'react-redux'
+import LabelledFlag from './labelled-flag'
+import DefaultFlag from './default-flag'
 
 const Read = ({ data = {} }) => {
   const sourceCode = useSelector(selectCode('USER'))
@@ -21,47 +19,35 @@ const Read = ({ data = {} }) => {
   return <Switch isChecked={data?.value} onChange={toggle} />
 }
 
-const Write = ({ questionCode, data, onSendAnswer, placeholderName: label, mandatory }) => {
-  const [isChecked, setIsChecked] = useState(false)
-  const clientId = apiConfig?.clientId
-  const handleToggle = () => {
-    onSendAnswer(`${!isChecked}`)
-    setIsChecked(isChecked => !isChecked)
+const Write = ({
+  questionCode,
+  data,
+  onSendAnswer,
+  placeholderName: label,
+  mandatory,
+  labelled,
+}) => {
+  if (labelled) {
+    return (
+      <LabelledFlag
+        questionCode={questionCode}
+        data={data}
+        onSendAnswer={onSendAnswer}
+        placeholderName={label}
+        mandatory={mandatory}
+      />
+    )
+  } else {
+    return (
+      <DefaultFlag
+        questionCode={questionCode}
+        data={data}
+        onSendAnswer={onSendAnswer}
+        placeholderName={label}
+        mandatory={mandatory}
+      />
+    )
   }
-
-  useEffect(() => {
-    if (isNullOrUndefinedOrEmpty(data?.value)) {
-      setIsChecked(false)
-      onSendAnswer(false)
-    } else {
-      setIsChecked(data?.value)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data?.value])
-
-  return (
-    <HStack spacing={5} justifyContent={'space-between'}>
-      <HStack>
-        <Text color="gray.700">{label}</Text>
-        {mandatory ? (
-          <Text as="span" color={'red.500'} ml={1}>
-            *
-          </Text>
-        ) : (
-          <></>
-        )}
-      </HStack>
-
-      <Box>
-        <Switch
-          colorScheme={equals(clientId)('lojing') ? 'orange' : 'primary'}
-          test-id={questionCode}
-          onChange={handleToggle}
-          value={isChecked}
-        />
-      </Box>
-    </HStack>
-  )
 }
 const Flag = {
   Write,
