+++ src/app/PCM/templates/tpl-detail-view/tpl-application-detail-view/checkLists.js
@@ -13,7 +13,13 @@ import { rejectUndefined } from 'utils/helpers/reject-undefined'
 
 const CheckLists = ({ code: passedQuestionCode, isStudent }) => {
   const CheckListItem = ({ parentCode, questionCode }) => {
-    return <Ask overrideComponent={'approve'} parentCode={parentCode} questionCode={questionCode} />
+    return (
+      <Ask
+        overrideComponent={'labelled_flag'}
+        parentCode={parentCode}
+        questionCode={questionCode}
+      />
+    )
   }
 
   /// Would be better to have the capabilities or backend handle this in the future, however
